Rendering, Saving, function callbacks
Pointers removed from scene header
Adding gameobjects no longer requires ID numbers
Shaders changed so depth is a float instead of a vec4

+++ src/graphics/Shader.h
@@ -26,7 +26,7 @@ public:
 private:
 	GLint shaderID;
 	std::ifstream shaderFile;
-	uint64_t shaderType;
+	uint32_t shaderType;
 };
 #endif // !SHADER_H
